Batch update (#238)
* feat: rewrite webdrivers to puppeteer

* feat: first step of multi-version build migration

* feat: second step of multi-version build migration

* feat: third step of multi-version build migration

* feat: fix tests for multi-version build migration

* chore: apply formaters

* fix: remove 3 non-deterministic bytes

* chore: keep debug info in wasm

* chore: cleanup devenv

* chore: cleanup preview1, fix address deserialization bug

* feat: change runners fs layout

* feat: fix tests

* feat: rework memory consumption to derived limiters

* chore: update github CI

* chore: pass CI

* Apply suggestions from code review


---------

+++ executor/src/wasi/vfs.rs
@@ -5,7 +5,7 @@ use genvm_common::*;
 
 pub struct FileContents {
     pub contents: util::SharedBytes,
-    pub pos: usize,
+    pub pos: u32,
 
     pub release_memory: bool,
 }
